requirements.txt
pytesseract
pillow
opencv-python
openai
python-dotenv
httpx[socks]